update conenction service

+++ index.php
@@ -70,14 +70,22 @@ if ($text && $chat_id) {
         $telegram->sendMessage(['chat_id' => $chat_id, 'parse_mode' => 'HTML', 'text' => $reply]);
         $connectionService->updateLastCommand($pdo, "/delete");
     } elseif ($lastCommand === "/delete") {
-        $deletedReasonText = $connectionService->getReasonById($pdo, $text);
+        if ((int)$text) {
+            $deletedReasonText = $connectionService->getReasonById($pdo, $text);
 
-        $connectionService->deleteReason($pdo, $text);
-        $connectionService->updateLastCommand($pdo);
+            $connectionService->deleteReason($pdo, $text);
+            $connectionService->updateLastCommand($pdo);
 
-        $reply = "Причина: $deletedReasonText была удалена. \nЖду, когда ты вернешь ее в список";
+            $reply = "Причина: $deletedReasonText была удалена. \nЖду, когда ты вернешь ее в список";
 
-        $telegram->sendMessage(['chat_id' => $chat_id, 'parse_mode' => 'HTML', 'text' => $reply]);
+            $telegram->sendMessage(['chat_id' => $chat_id, 'parse_mode' => 'HTML', 'text' => $reply]);
+        } else {
+            $connectionService->updateLastCommand($pdo);
+            $reply = "Циферку вводи. Ты справишься, я уверен.";
+
+            $telegram->sendMessage(['chat_id' => $chat_id, 'parse_mode' => 'HTML', 'text' => $reply]);
+        }
+
     } elseif ($text === "/show") {
         $reply = "Вспомни все свои грехи";
